Create sov-app-template
