+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
 	"demo-app",
 	"sdk",
 	"first-read-last-write-cache",
+	"sov-modules/sov-app-template",
 	"sov-modules/sov-modules-macros",
 	"sov-modules/sov-state",
 	"sov-modules/sov-modules-api",
@@ -30,6 +31,7 @@ rust-version = "1.67"
 # Internal dependencies
 sovereign-sdk = {path= "sdk" }
 first-read-last-write-cache = {path = "first-read-last-write-cache"}
+sov-app-template = { path = "sov-modules/sov-app-template" }
 election = { path = "sov-modules/sov-modules-impl/examples/election" }
 value-setter = { path = "sov-modules/sov-modules-impl/examples/value-setter" }
 accounts = { path = "sov-modules/sov-modules-impl/accounts" }
