+++ demo-app/src/data_generation.rs
@@ -1,9 +1,7 @@
-use crate::{
-    runtime::Runtime,
-    tx_verifier::{RawTx, Transaction},
-};
+use crate::runtime::Runtime;
 use borsh::BorshSerialize;
-use sov_modules_api::mocks::{MockContext, MockPublicKey, MockSignature};
+use sov_app_template::{RawTx, Transaction};
+use sov_modules_api::mocks::{MockContext, MockPublicKey};
 use sov_modules_api::PublicKey;
 
 pub(crate) fn simulate_da() -> Vec<RawTx> {
@@ -108,18 +106,5 @@ impl QueryGenerator {
     pub(crate) fn generate_query_value_setter_message() -> Vec<u8> {
         let query_message = value_setter::query::QueryMessage::GetValue;
         Runtime::<MockContext>::encode_value_setter_query(query_message)
-    }
-}
-
-impl Transaction<MockContext> {
-    pub fn new(msg: Vec<u8>, pub_key: MockPublicKey, nonce: u64) -> Self {
-        Self {
-            signature: MockSignature {
-                msg_sig: Vec::default(),
-            },
-            runtime_msg: msg,
-            pub_key,
-            nonce,
-        }
     }
 }
