+++ sov-modules/sov-app-template/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "demo-app"
+name = "sov-app-template"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,14 +8,10 @@ edition = "2021"
 [dependencies]
 anyhow = { workspace = true }
 sovereign-sdk = { workspace = true, features = ["mocks"] }
-election = { workspace = true }
-sov-app-template = { workspace = true }
-value-setter = { workspace = true }
-accounts = { workspace = true }
 sov-state = { workspace = true, features = ["temp"] }
 sovereign-db = { workspace = true }
-sov-modules-api = { workspace = true, features = ["mocks"] }
 sov-modules-macros = { workspace = true }
+sov-modules-api = { workspace = true }
 borsh = { workspace = true }
 serde_json = { workspace = true }
 serde = { workspace = true }
